integrations/joytag/requirements-server.txt
torch>=2.0.1,<3.0.0
torchvision>=0.15.2
einops>=0.7.0
safetensors>=0.4.1
pillow>=9.4.0
flask>=2.3.0
requests>=2.31.0
transformers>=4.36.2
